flowiseQA/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
requests==2.32.3
SQLAlchemy==2.0.35
pymysql==1.1.1
python-dotenv==1.0.1
pydantic==2.8.2
cryptography==43.0.1